requirements.txt
# requirements.txt
requests
beautifulsoup4
google-generativeai
python-dotenv
pyyaml
jinja2
playwright
lxml